[SHELL32] split dialogs/folder_options.cpp
