+++ dll/win32/shell32/precomp.h
@@ -86,6 +86,7 @@
 #include "shellmenu/CMergedFolder.h"
 #include "shellmenu/shellmenu.h"
 #include "CUserNotification.h"
+#include "dialogs/folder_options.h"
 
 #include <wine/debug.h>
 #include <wine/unicode.h>
